Added a catch for installation problems

+++ omz_install.sh
@@ -2,10 +2,12 @@
 if [ -t 1 ]; then
   BLUE=$(printf '\033[34m')
   GREEN=$(printf '\033[32m')
+  RED=$(printf '\033[31m')
   RESET=$(printf '\033[m')
 else
   BLUE=""
   GREEN=""
+  RED=""
   RESET=""
 fi
 
@@ -15,34 +17,39 @@ cd $HOME
 echo "${BLUE}Downloading and installing Oh My Zsh...${RESET}"
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
 
-# Download and install Powerline fonts
-echo "${BLUE}Downloading and installing Powerline fonts...${RESET}"
-git clone https://github.com/powerline/fonts.git --depth=1
-cd fonts
-./install.sh
-cd ..
-rm -rf fonts
+# Proceed if Oh My Zsh installed successfully 
+if [ $? -eq 0 ]; then
+  # Download and install Powerline fonts
+  echo "${BLUE}Downloading and installing Powerline fonts...${RESET}"
+  git clone https://github.com/powerline/fonts.git --depth=1
+  cd fonts
+  ./install.sh
+  cd ..
+  rm -rf fonts
 
-echo "${BLUE}Correcting your zsh configuration...${RESET}"
-cd $HOME  
+  echo "${BLUE}Correcting your zsh configuration...${RESET}"
+  cd $HOME  
 
-# Save OMZ's full version of .zshrc file
-mv .zshrc .zshrc.oh-my-zsh
+  # Save OMZ's full version of .zshrc file
+  mv .zshrc .zshrc.oh-my-zsh
 
-# Create a new .zshrc file with only essential OMZ commands
-echo "# Path to your Oh My Zsh installation.
-export ZSH=\"$HOME/.oh-my-zsh\"
+  # Create a new .zshrc file with only essential OMZ commands
+  echo "# Path to your Oh My Zsh installation.
+  export ZSH=\"$HOME/.oh-my-zsh\"
 
-# Set name of the theme to load (See https://github.com/ohmyzsh/ohmyzsh/wiki/Themes)
-ZSH_THEME=\"dpoggi\"
+  # Set name of the theme to load (See https://github.com/ohmyzsh/ohmyzsh/wiki/Themes)
+  ZSH_THEME=\"dpoggi\"
 
-# Activate git plugin
-plugins=(git)
+  # Activate git plugin
+  plugins=(git)
 
-# Activate Oh My Zsh
-source \$ZSH/oh-my-zsh.sh \n" > .zshrc
+  # Activate Oh My Zsh
+  source \$ZSH/oh-my-zsh.sh \n" > .zshrc
 
-# Add content from initial .zshrc file to the current one
-cat .zshrc.pre-oh-my-zsh >> .zshrc
+  # Add content from initial .zshrc file to the current one
+  cat .zshrc.pre-oh-my-zsh >> .zshrc
 
-echo "${GREEN}All done! Reopen terminal and enjoy your new colorful shell!${RESET}"
+  echo "${GREEN}All done! Reopen terminal and enjoy your new colorful shell!${RESET}"
+else
+  echo "${RED}Something went wrong. Oh My Zsh cannot be installed.${RESET}"
+fi
